AG-15107 - Cleanup high-freq capable series.
